+++ src/components/StatusMenu/StatusBarButton.vue
@@ -13,17 +13,17 @@ export default {
 <style scoped>
 .statusBarButton {
 	position: relative;
+	display: flex;
+	align-items: center;
+	justify-content: center;
 	
 	height: 40px;
+	width: 40px;
 
 	border-radius: 20px;
 	border: 3px solid #E0E0E0;
-	padding: 0 15px;
 
 	color: #828282;
-	font-size: 18px;
-	line-height: 36px;
-	font-weight: 700;
 
 	transition: all 0.2s;
 }