+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 ---
 ### Beta 0.3
 + Add changelog
+* Replace status bar button text with icons